backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
flask-jwt-extended==4.7.1
langgraph==0.4.1
langchain-anthropic==0.3.12
langchain-core>=0.3.53,<1.0.0
python-dotenv==1.1.0
gunicorn==23.0.0
xmltodict==0.14.1
httpx==0.28.1
pytest==8.3.5
pytest-flask==1.3.0
werkzeug==3.1.3
